refactor(client): extract attachment and mod tests to sibling files
Moves the inline #[cfg(test)] mod tests blocks from attachment.rs
and mod.rs into sibling attachment_tests.rs and mod_tests.rs files,
matching the existing pattern at comment.rs / comment_tests.rs.
Brings the test fixtures under the existing CPD exclusion in
sonar-project.properties (which excludes **/*_tests.rs from
copy-paste detection). Resolves the SonarCloud quality-gate failure
on PR #137 (13.9% duplicated lines on new code, threshold 3%) caused
by repetitive wiremock JSON fixtures that are categorically test-side
boilerplate, not production duplication.
